[REEF-664] REEF-IO test and main packages do not match
- Move tests that were in org.apache.reef.services to org.apache.reef.io.
- Remove a duplicate class and an unused class

JIRA:
  [REEF-664](https://issues.apache.org/jira/browse/REEF-664)

Pull Request:
  Closes #
