+++ lang/java/reef-io/src/test/java/org/apache/reef/io/network/naming/LocalNameResolverTest.java
@@ -16,10 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.reef.services.network;
+package org.apache.reef.io.network.naming;
 
-import org.apache.reef.io.network.naming.LocalNameResolverConfiguration;
-import org.apache.reef.io.network.naming.NameResolver;
 import org.apache.reef.io.network.naming.exception.NamingException;
 import org.apache.reef.io.network.util.StringIdentifierFactory;
 import org.apache.reef.tang.Tang;
