+++ lang/java/reef-io/src/test/java/org/apache/reef/io/network/naming/NameClientTest.java
@@ -16,9 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.reef.services.network;
+package org.apache.reef.io.network.naming;
 
-import org.apache.reef.io.network.naming.*;
 import org.apache.reef.io.network.naming.exception.NamingException;
 import org.apache.reef.io.network.naming.parameters.NameResolverRetryCount;
 import org.apache.reef.io.network.naming.parameters.NameResolverRetryTimeout;
