+++ lang/java/reef-io/src/test/java/org/apache/reef/io/network/util/NetworkMessagingTestService.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.reef.services.network.util;
+package org.apache.reef.io.network.util;
 
 import org.apache.reef.exception.evaluator.NetworkException;
 import org.apache.reef.io.network.Connection;
@@ -41,24 +41,20 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-// TODO[JIRA REEF-637] Remove the deprecated class.
 /**
- * Helper class for DeprecatedNetworkConnectionService test, deprecated in 0.13.
+ * Helper class for NetworkConnectionService test.
  */
-@Deprecated
-public final class DeprecatedNetworkMessagingTestService implements AutoCloseable {
-  private static final Logger LOG = Logger.getLogger(DeprecatedNetworkMessagingTestService.class.getName());
+public final class NetworkMessagingTestService implements AutoCloseable {
+  private static final Logger LOG = Logger.getLogger(NetworkMessagingTestService.class.getName());
 
   private final IdentifierFactory factory;
   private final NetworkConnectionService receiverNetworkConnService;
   private final NetworkConnectionService senderNetworkConnService;
-  private final String receiver;
-  private final String sender;
   private final NameServer nameServer;
   private final NameResolver receiverResolver;
   private final NameResolver senderResolver;
 
-  public DeprecatedNetworkMessagingTestService(final String localAddress) throws InjectionException {
+  public NetworkMessagingTestService(final String localAddress) throws InjectionException {
     // name server
     final Injector injector = Tang.Factory.getTang().newInjector();
     this.nameServer = injector.getInstance(NameServer.class);
@@ -69,34 +65,36 @@ public final class DeprecatedNetworkMessagingTestService implements AutoCloseabl
 
     LOG.log(Level.FINEST, "=== Test network connection service receiver start");
     // network service for receiver
-    this.receiver = "receiver";
     final Injector injectorReceiver = injector.forkInjector(netConf);
     this.receiverNetworkConnService = injectorReceiver.getInstance(NetworkConnectionService.class);
     this.receiverResolver = injectorReceiver.getInstance(NameResolver.class);
     this.factory = injectorReceiver.getNamedInstance(NetworkConnectionServiceIdFactory.class);
-    this.receiverNetworkConnService.registerId(this.factory.getNewInstance(receiver));
 
     // network service for sender
-    this.sender = "sender";
     LOG.log(Level.FINEST, "=== Test network connection service sender start");
     final Injector injectorSender = injector.forkInjector(netConf);
     senderNetworkConnService = injectorSender.getInstance(NetworkConnectionService.class);
-    senderNetworkConnService.registerId(this.factory.getNewInstance(this.sender));
     this.senderResolver = injectorSender.getInstance(NameResolver.class);
   }
 
   public <T> void registerTestConnectionFactory(final Identifier connFactoryId,
                                                 final int numMessages, final Monitor monitor,
                                                 final Codec<T> codec) throws NetworkException {
+    final Identifier receiverEndPointId = factory.getNewInstance("receiver");
+    final Identifier senderEndPointId = factory.getNewInstance("sender");
     receiverNetworkConnService.registerConnectionFactory(connFactoryId, codec,
-        new MessageHandler<T>(monitor, numMessages), new TestListener<T>());
+        new MessageHandler<T>(monitor, numMessages, senderEndPointId, receiverEndPointId),
+        new TestListener<T>(), receiverEndPointId);
     senderNetworkConnService.registerConnectionFactory(connFactoryId, codec,
-        new MessageHandler<T>(monitor, numMessages), new TestListener<T>());
+        new MessageHandler<T>(monitor, numMessages, receiverEndPointId, senderEndPointId),
+        new TestListener<T>(), senderEndPointId);
   }
 
   public <T> Connection<T> getConnectionFromSenderToReceiver(final Identifier connFactoryId) {
-    final Identifier destId = factory.getNewInstance(receiver);
-    return (Connection<T>)senderNetworkConnService.getConnectionFactory(connFactoryId).newConnection(destId);
+    final Identifier receiverEndPointId = factory.getNewInstance("receiver");
+    return (Connection<T>)senderNetworkConnService
+        .getConnectionFactory(connFactoryId)
+        .newConnection(receiverEndPointId);
   }
 
   public void close() throws Exception {
@@ -110,12 +108,18 @@ public final class DeprecatedNetworkMessagingTestService implements AutoCloseabl
   public static final class MessageHandler<T> implements EventHandler<Message<T>> {
     private final int expected;
     private final Monitor monitor;
+    private final Identifier expectedSrcId;
+    private final Identifier expectedDestId;
     private AtomicInteger count = new AtomicInteger(0);
 
     public MessageHandler(final Monitor monitor,
-                          final int expected) {
+                          final int expected,
+                          final Identifier expectedSrcId,
+                          final Identifier expectedDestId) {
       this.monitor = monitor;
       this.expected = expected;
+      this.expectedSrcId = expectedSrcId;
+      this.expectedDestId = expectedDestId;
     }
 
     @Override
@@ -129,6 +133,9 @@ public final class DeprecatedNetworkMessagingTestService implements AutoCloseabl
       for (final T obj : value.getData()) {
         LOG.log(Level.FINE, "OUT: data: {0}", obj);
       }
+
+      assert value.getSrcId().equals(expectedSrcId);
+      assert value.getDestId().equals(expectedDestId);
 
       if (count.get() == expected) {
         monitor.mnotify();
