+++ lang/java/reef-io/src/test/java/org/apache/reef/io/storage/ExternalMapTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.reef.services.storage;
+package org.apache.reef.io.storage;
 
 import org.apache.reef.io.ExternalMap;
 import org.apache.reef.io.serialization.Codec;
