+++ lang/java/reef-io/src/test/java/org/apache/reef/io/storage/package-info.java
@@ -19,4 +19,4 @@
 /**
  * TODO: Document.
  */
-package org.apache.reef.services.network;
+package org.apache.reef.io.storage;
